requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
certifi==2024.12.14
    # via requests
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.4.1
    # via requests
click==8.1.7
    # via uvicorn
django==5.1.4
    # via
    #   -r requirements.in
    #   django-ckeditor
    #   django-filter
    #   django-js-asset
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-ckeditor==6.7.2
    # via -r requirements.in
django-filter==24.3
    # via -r requirements.in
django-js-asset==3.0.1
    # via django-ckeditor
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.4.0
    # via -r requirements.in
drf-yasg==1.21.8
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-yasg
numpy==2.2.2
    # via pandas
packaging==24.2
    # via drf-yasg
pandas==2.2.3
    # via -r requirements.in
pillow==11.0.0
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via
    #   drf-yasg
    #   pandas
pyyaml==6.0.2
    # via drf-yasg
requests==2.32.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.2
    # via django
tzdata==2025.1
    # via pandas
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.3.0
    # via requests
uvicorn==0.32.1
    # via -r requirements.in
